+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fix TagVersion detection and require tagged build for publishing
 
 ## [0.3.2] / 2023-11-10
 - Support test results for multiple target frameworks